# storage/engine: reimplement MVCC{Scan,Get} in C++