--- a/pkg/roachpb/data.go
+++ b/pkg/roachpb/data.go
@@ -274,6 +274,25 @@ func (v Value) Verify(key []byte) error {
 	return nil
 }
 
+// VerifyValues verifies the checksums for the key/value pairs. The first error
+// encountered is returned.
+func VerifyValues(kvs []KeyValue) error {
+	crc := crc32Pool.Get().(hash.Hash32)
+	var err error
+	for i := range kvs {
+		kv := &kvs[i]
+		if sum := kv.Value.checksum(); sum != 0 {
+			if computedSum := computeChecksum(kv.Key, kv.Value.RawBytes, crc); computedSum != sum {
+				err = fmt.Errorf("%s: invalid checksum (%x) value [% x]",
+					kv.Key, computedSum, kv.Value.RawBytes)
+				break
+			}
+		}
+	}
+	crc32Pool.Put(crc)
+	return err
+}
+
 // ShallowClone returns a shallow clone of the receiver.
 func (v *Value) ShallowClone() *Value {
 	if v == nil {
@@ -562,28 +581,33 @@ var crc32Pool = sync.Pool{
 	},
 }
 
-// computeChecksum computes a checksum based on the provided key and
-// the contents of the value.
-func (v Value) computeChecksum(key []byte) uint32 {
-	if len(v.RawBytes) < headerSize {
+func computeChecksum(key, rawBytes []byte, crc hash.Hash32) uint32 {
+	if len(rawBytes) < headerSize {
 		return 0
 	}
-	crc := crc32Pool.Get().(hash.Hash32)
 	if _, err := crc.Write(key); err != nil {
 		panic(err)
 	}
-	if _, err := crc.Write(v.RawBytes[checksumSize:]); err != nil {
+	if _, err := crc.Write(rawBytes[checksumSize:]); err != nil {
 		panic(err)
 	}
 	sum := crc.Sum32()
 	crc.Reset()
-	crc32Pool.Put(crc)
 	// We reserved the value 0 (checksumUninitialized) to indicate that a checksum
 	// has not been initialized. This reservation is accomplished by folding a
 	// computed checksum of 0 to the value 1.
 	if sum == checksumUninitialized {
 		return 1
 	}
+	return sum
+}
+
+// computeChecksum computes a checksum based on the provided key and
+// the contents of the value.
+func (v Value) computeChecksum(key []byte) uint32 {
+	crc := crc32Pool.Get().(hash.Hash32)
+	sum := computeChecksum(key, v.RawBytes, crc)
+	crc32Pool.Put(crc)
 	return sum
 }