--- a/pkg/storage/engine/rocksdb.go
+++ b/pkg/storage/engine/rocksdb.go
@@ -2005,9 +2005,6 @@ func (r *rocksDBIterator) MVCCGet(
 		Timestamp: mvccKey.Timestamp,
 	}
 	copy(value.RawBytes, rawValue)
-	if err := value.Verify(mvccKey.Key); err != nil {
-		return nil, nil, err
-	}
 	return value, intents, nil
 }