Reimplement `MVCC{Scan,Get}` in C++. The general strategy is for the C++
version of `MVCC{Scan,Get}` to return key/values in the RocksDB batch
repr format which we can iterate over in Go using
`RocksDBBatchReader`. Intents are returned in a separate batch which Go
code translates into either a `WriteIntentError` or returns the intents
to the caller (depending on the `consistent` and `txn` settings).

`SELECT COUNT(*) FROM test.vk` where `test.kv` contains 5m rows reduced
from 4.3s to 2.3s (a 46% speedup) on a local single-node cluster.

The `versions=100` benchmarks have regressed because of the optimization
to call `iter->Next()` up to 10 times vs only a single time in the old
code (see `DBIterNext`). This hurts the `versions=100` benchmarks, but
significantly helps the `versions=2` and `versions=10` benchmarks.

The C++ version of `MVCCGet` is implemented in terms of `MVCCScan` which
likely accounts for the performance drop in the `MVCCGet` benchmarks. It
should be possible to recover this performance difference with
specialization of the C++ code.

```
name                                                   old time/op    new time/op     delta
MVCCScan_RocksDB/rows=1/versions=1/valueSize=8-8         5.96µs ± 2%     5.81µs ± 2%    -2.51%  (p=0.000 n=9+10)
MVCCScan_RocksDB/rows=1/versions=2/valueSize=8-8         6.95µs ± 2%     5.86µs ± 2%   -15.60%  (p=0.000 n=10+10)
MVCCScan_RocksDB/rows=1/versions=10/valueSize=8-8        8.27µs ± 1%     6.21µs ± 2%   -24.94%  (p=0.000 n=8+10)
MVCCScan_RocksDB/rows=1/versions=100/valueSize=8-8       9.44µs ± 2%     7.59µs ± 2%   -19.61%  (p=0.000 n=8+9)
MVCCScan_RocksDB/rows=10/versions=1/valueSize=8-8        12.7µs ± 5%      9.3µs ± 2%   -26.83%  (p=0.000 n=9+9)
MVCCScan_RocksDB/rows=10/versions=2/valueSize=8-8        20.0µs ± 1%     10.8µs ±16%   -45.89%  (p=0.000 n=10+10)
MVCCScan_RocksDB/rows=10/versions=10/valueSize=8-8       31.2µs ± 2%     14.1µs ± 2%   -54.66%  (p=0.000 n=9+8)
MVCCScan_RocksDB/rows=10/versions=100/valueSize=8-8      37.1µs ± 1%     38.8µs ± 1%    +4.63%  (p=0.000 n=9+8)
MVCCScan_RocksDB/rows=100/versions=1/valueSize=8-8       71.6µs ± 1%     39.9µs ± 2%   -44.29%  (p=0.000 n=10+10)
MVCCScan_RocksDB/rows=100/versions=2/valueSize=8-8        139µs ± 2%       47µs ± 2%   -66.27%  (p=0.000 n=10+10)
MVCCScan_RocksDB/rows=100/versions=10/valueSize=8-8       246µs ± 5%       85µs ± 2%   -65.32%  (p=0.000 n=10+10)
MVCCScan_RocksDB/rows=100/versions=100/valueSize=8-8      300µs ± 1%      335µs ± 2%   +11.77%  (p=0.000 n=8+10)
MVCCScan_RocksDB/rows=1000/versions=1/valueSize=8-8       635µs ± 1%      335µs ± 1%   -47.28%  (p=0.000 n=9+10)
MVCCScan_RocksDB/rows=1000/versions=2/valueSize=8-8      1.30ms ± 2%     0.40ms ± 1%   -69.41%  (p=0.000 n=10+10)
MVCCScan_RocksDB/rows=1000/versions=10/valueSize=8-8     2.39ms ± 6%     0.80ms ±10%   -66.37%  (p=0.000 n=10+10)
MVCCScan_RocksDB/rows=1000/versions=100/valueSize=8-8    3.12ms ± 1%     3.27ms ± 3%    +4.61%  (p=0.000 n=7+9)
MVCCGet_RocksDB/versions=1/valueSize=8-8                 5.63µs ± 3%     5.93µs ± 3%    +5.49%  (p=0.000 n=9+10)
MVCCGet_RocksDB/versions=10/valueSize=8-8                6.73µs ±13%     6.51µs ± 2%      ~     (p=0.190 n=10+10)
MVCCGet_RocksDB/versions=100/valueSize=8-8               11.1µs ± 8%     11.5µs ± 3%      ~     (p=0.089 n=10+10)
```

Fixes #21334

Release note (performance improvement): Reimplement low-level scan
operations in C++, doubling performance for scans of contiguous
keys/rows.