@petermattis 2018-01-11
Sending this out for review because all of the tests pass and the performance improvement is significant, but there is more work to be done. In particular, I haven't implemented `MVCCReverseScan` yet and I want to fix the `MVCCGet` performance regression. Once everything is performing to our satisfaction, there is a chunk of Go code that we'll be able to cleanup such as reimplementing `MVCCIterate` in terms of `MVCCScan`, and deleting `mvccGetInternal` and associated routines. 

Lastly, I need to document the crap out of all of the cases in `mvccScanner`. I'm planning to do that ASAP while they are all fresh in my mind.