@petermattis 2018-01-11
Review status: 0 of 10 files reviewed at latest revision, 1 unresolved discussion.

---

*[pkg/storage/engine/engine.go, line 117 at r1](https://reviewable.io:443/reviews/cockroachdb/cockroach/21395#-L2Y5M4mXuMZzxN5IUDm:-L2Y5M4mXuMZzxN5IUDn:b-phsvb0) ([raw file](https://github.com/cockroachdb/cockroach/blob/821514fcee1bb1d2877d3175d2a8c33ef95f8a89/pkg/storage/engine/engine.go#L117)):*
> ```Go
> 	MVCCScan(start, end roachpb.Key, max int64, timestamp hlc.Timestamp,
> 		txn *roachpb.Transaction, consistent, reverse bool,
> 	) (kvs []byte, intents []byte, err error)
> ```

We could have `MVCCScan` return `[]roachpb.KeyValue` and `[]roachpb.Intent`, but returning batch repr ties in with #18913 which we should take another look at.

---


*Comments from [Reviewable](https://reviewable.io:443/reviews/cockroachdb/cockroach/21395)*
<!-- Sent from Reviewable.io -->